refactor: add parametrized PaginatedResponseSchema and move pagination schemas to schemas/common
Introduce a reusable PaginatedResponseSchema(dataSchema) factory and relocate
the pagination schemas out of types/ into schemas/common/pagination, replacing
the hand-rolled { data, meta } response structs across list hooks.
