+++ src/hooks/api/businesses/[business-id]/customers/useListCustomers.ts
@@ -2,7 +2,7 @@ import { useCallback } from 'react'
 import { Schema } from 'effect'
 import useSWRInfinite from 'swr/infinite'
 
-import { PaginatedResponseMetaSchema } from '@internal-types/utility/pagination'
+import { PaginatedResponseSchema } from '@schemas/common/pagination'
 import { type Customer, CustomerSchema } from '@schemas/customer'
 import { get } from '@utils/api/authenticatedHttp'
 import { toDefinedSearchParameters } from '@utils/request/toDefinedSearchParameters'
@@ -13,12 +13,7 @@ import { usePreserveInfiniteSize } from '@utils/swr/usePreserveInfiniteSize'
 import { useAuth } from '@hooks/utils/auth/useAuth'
 import { useLayerContext } from '@contexts/LayerContext/LayerContext'
 
-const ListCustomersRawResultSchema = Schema.Struct({
-  data: Schema.Array(CustomerSchema),
-  meta: Schema.Struct({
-    pagination: PaginatedResponseMetaSchema,
-  }),
-})
+const ListCustomersRawResultSchema = PaginatedResponseSchema(CustomerSchema)
 type ListCustomersRawResult = typeof ListCustomersRawResultSchema.Type
 
 type ListCustomersBaseParams = {
