+++ src/hooks/api/businesses/[business-id]/mileage/trips/useListTrips.tsx
@@ -2,7 +2,7 @@ import { useCallback } from 'react'
 import { Schema } from 'effect'
 import useSWRInfinite from 'swr/infinite'
 
-import { PaginatedResponseMetaSchema } from '@internal-types/utility/pagination'
+import { PaginatedResponseSchema } from '@schemas/common/pagination'
 import { type Trip, TripSchema } from '@schemas/trip'
 import { get } from '@utils/api/authenticatedHttp'
 import { toDefinedSearchParameters } from '@utils/request/toDefinedSearchParameters'
@@ -22,12 +22,7 @@ export type ListTripsFilterParams = {
   year?: number
 }
 
-const ListTripsResponseSchema = Schema.Struct({
-  data: Schema.Array(TripSchema),
-  meta: Schema.Struct({
-    pagination: PaginatedResponseMetaSchema,
-  }),
-})
+const ListTripsResponseSchema = PaginatedResponseSchema(TripSchema)
 type ListTripsResponse = typeof ListTripsResponseSchema.Type
 
 function keyLoader(
